requirements.txt
# Pramana API Dependencies

# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Storage
pyarrow>=15.0.0
b2sdk>=2.0.0

# HTTP client
httpx>=0.26.0

# Validation
pydantic>=2.5.0

# Authentication
PyJWT>=2.8.0

# Multipart forms
python-multipart>=0.0.9
